+++ addOns/authhelper/src/main/java/org/zaproxy/addon/authhelper/AuthenticationDiagnostics.java
@@ -29,7 +29,6 @@ import java.util.Objects;
 import java.util.function.Supplier;
 import javax.jdo.PersistenceManager;
 import javax.jdo.Transaction;
-import org.apache.commons.httpclient.URIException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.openqa.selenium.By;
@@ -56,7 +55,7 @@ import org.zaproxy.addon.authhelper.internal.db.DiagnosticWebElement;
 import org.zaproxy.addon.authhelper.internal.db.DiagnosticWebElement.SelectorType;
 import org.zaproxy.addon.authhelper.internal.db.TableJdo;
 import org.zaproxy.zap.extension.zest.ZestZapUtils;
-import org.zaproxy.zap.model.SessionStructure;
+import org.zaproxy.zap.model.Context;
 import org.zaproxy.zap.network.HttpSenderListener;
 import org.zaproxy.zest.core.v1.ZestClientElement;
 import org.zaproxy.zest.core.v1.ZestClientElementClear;
@@ -75,7 +74,7 @@ public class AuthenticationDiagnostics implements AutoCloseable {
     private static final List<DiagnosticDataProvider> diagnosticDataProviders =
             Collections.synchronizedList(new ArrayList<>());
 
-    private static final List<DomainAccessedConsumer> domainAccessedConsumers =
+    private static final List<MessageAccessedConsumer> messageAccessedConsumers =
             Collections.synchronizedList(new ArrayList<>());
 
     private static final String ELEMENT_SELECTOR_SCRIPT =
@@ -192,7 +191,7 @@ function getSelector(element, documentElement) {
 return getSelector(arguments[0], document)
 """;
 
-    private final HttpSenderListener domainAccessedListener;
+    private final HttpSenderListener messageAccessedListener;
 
     private final boolean enabled;
 
@@ -214,8 +213,10 @@ return getSelector(arguments[0], document)
             String script) {
         this.enabled = enabled;
 
-        domainAccessedListener =
+        messageAccessedListener =
                 new HttpSenderListener() {
+
+                    private Context ctx = Model.getSingleton().getSession().getContext(context);
 
                     @Override
                     public void onHttpResponseReceive(
@@ -227,12 +228,7 @@ return getSelector(arguments[0], document)
                     @Override
                     public void onHttpRequestSend(
                             HttpMessage msg, int initiator, HttpSender sender) {
-                        try {
-                            String domain = SessionStructure.getHostName(msg);
-                            domainAccessedConsumers.forEach(e -> e.domainAccessed(domain));
-                        } catch (URIException ignore) {
-                            // Nothing to do.
-                        }
+                        messageAccessedConsumers.forEach(e -> e.messageAccessed(ctx, msg));
                     }
 
                     @Override
@@ -240,7 +236,7 @@ return getSelector(arguments[0], document)
                         return 0;
                     }
                 };
-        HttpSender.addListener(domainAccessedListener);
+        HttpSender.addListener(messageAccessedListener);
 
         if (!enabled) {
             return;
@@ -534,7 +530,7 @@ return getSelector(arguments[0], document)
 
     @Override
     public void close() {
-        HttpSender.removeListener(domainAccessedListener);
+        HttpSender.removeListener(messageAccessedListener);
 
         if (!enabled) {
             return;
@@ -612,19 +608,19 @@ return getSelector(arguments[0], document)
         }
     }
 
-    public static void addDomainsAccessedConsumer(DomainAccessedConsumer consumer) {
+    public static void addMessageAccessedConsumer(MessageAccessedConsumer consumer) {
         Objects.requireNonNull(consumer);
-        domainAccessedConsumers.add(consumer);
+        messageAccessedConsumers.add(consumer);
     }
 
-    public static void removeDomainsAccessedConsumer(DomainAccessedConsumer consumer) {
+    public static void removeMessageAccessedConsumer(MessageAccessedConsumer consumer) {
         Objects.requireNonNull(consumer);
-        domainAccessedConsumers.remove(consumer);
+        messageAccessedConsumers.remove(consumer);
     }
 
-    public interface DomainAccessedConsumer {
+    public interface MessageAccessedConsumer {
 
-        void domainAccessed(String domain);
+        void messageAccessed(Context ctx, HttpMessage message);
     }
 
     public static void addDiagnosticDataProvider(DiagnosticDataProvider provider) {