+++ addOns/authhelper/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Added
 - Handle account selection and TOTP step in Microsoft login.
+- Allow to include domains completely and partially out of scope in the Authentication Report.
 
 ### Changed
 - Fail the Microsoft login if not able to perform all the expected steps.